requirements.txt
flet
multitasking
psutil
pycryptodome
pyhocon
pyperclip
requests
retry
